Add capability to set VLANs on uplink interfaces
Configuration is driven by an optional configuration option.
If set to true the uplink PF (or bond if its part of one) has the
same VLANs added to it as the VF being added.

When removing VF this same configuration will remove vlans from the
uplink, but only if no other VF are on the bridge with the same vlans.

Create new netlink support functions to delete interfaces from a trunk
and get a complete list of bridge interfaces and their netlink vlan info